requirements.txt
Flask==2.3.3
requests==2.31.0
pushbullet.py==0.12.0
python-dotenv==1.0.0
APScheduler==3.10.4
